+++ lib/backend/repository/interfaces/consumerRepositoryInterface.dart
@@ -1,9 +1,7 @@
-import '../models/consumer.dart';
+import 'package:uplace/backend/repository/interfaces/baseRepository.dart';
+import 'package:uplace/backend/repository/models/consumer.dart';
 
-abstract class ConsumerRepositoryInterface {
-  bool configured = false;
-
-  void configure();
+abstract class ConsumerRepositoryInterface extends BaseRepository {
   Future<bool> createConsumer(String name, DateTime birthDate, String userUID);
   Future<Consumer?> getConsumerById(consumerId);
 }
